% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathifiles.R
\name{add_col_from_workset}
\alias{add_col_from_workset}
\title{Adds a column to a hathifile from a workset}
\usage{
add_col_from_workset(hathi_file, workset, col_name)
}
\arguments{
\item{hathi_file}{The hathifile in memory, typically loaded by
\link{load_raw_hathifile} (and perhaps filtered etc.). Must contain an "htid"
column.}

\item{workset}{A workset of htids, generated by \link{workset_builder} from \href{https://solr2.htrc.illinois.edu/solr-ef/}{Hathi Trust's SOLR endpoint}.}

\item{col_name}{A name for the new created column. If missing, it will
attempt to generate a column name from the "query" attribute of the workset
(\code{attr(workset, "query")}); note that this may result in unwieldy and
complex column names.}
}
\value{
The original hathi_file, plus an added column based on the workset
named \code{col_name} and containing the number of pages where the query from
the workset is found.
}
\description{
Adds a column to a hathifile (or a sample of a hathifile, containing an
"htid" column) from a workset generated by \link{workset_builder}. This is
basically a convenience wrapper, with some checks, of \link[dplyr:mutate-joins]{dplyr::left_join}. One
could just as well download the workset and join to the hathifile directly.
}
\examples{
\dontrun{
hf <- load_raw_hathifile(cols = c("htid", "rights_date_used", "title", "author"))
res <- workset_builder("democracy") # All pages that mention democracy. 2M volumes.
hf <- hf \%>\% add_col_from_workset(res, "num_mentions_dem")
}
}
