% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathifiles.R
\name{add_drama_indicator}
\alias{add_drama_indicator}
\title{Adds a "drama" indicator to a hathifile}
\usage{
add_drama_indicator(hathi_file)
}
\arguments{
\item{hathi_file}{The hathifile, typically loaded by \link{load_raw_hathifile}.}
}
\value{
The hathifile with two new columns, \code{drama_prob} (the probability
that more than 80\% of the pages in the volume assigned to "fiction" have
been correctly classified.) and \code{drama_prop} (the proportion of pages in
the volume classified as fiction).
}
\description{
Adds a drama indicator based on the genre classification in Ted Underwood,
Boris Capitanu, Peter Organisciak, Sayan Bhattacharyya, Loretta Auvil,
Colleen Fallaw, J. Stephen Downie (2015). Word Frequencies in
English-Language Literature, 1700-1922 (0.2)
\href{http://data.analytics.hathitrust.org/genre/drama_metadata.csv}{Dataset}.
HathiTrust Research Center. doi:10.13012/J8JW8BSJ. See the documentation of
the drama dataset in \link{drama}. This is just a convenient wrapper; one
could easily just download the original drama metadata from
http://data.analytics.hathitrust.org/genre/drama_metadata.csv, select the
required columns, and join to the hathifile.
}
\examples{
\dontrun{
hf <- load_raw_hathifile(cols = c("htid", "rights_date_used", "title", "author")) \%>\%
  add_drama_indicator()
}
}
