% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathifiles.R
\name{add_imputed_date}
\alias{add_imputed_date}
\title{Add imputed date}
\usage{
add_imputed_date(hathi_file)
}
\arguments{
\item{hathi_file}{The hathifile in memory, typically loaded by
\link{load_raw_hathifile} (and perhaps filtered etc.). Must contain an "imprint"
column.}
}
\value{
The hathifile with added \verb{imputed date} column.
}
\description{
Adds an 'imputed' date of publication column based on the \code{imprint} column of
the hathifile. This function checks for a year in the \code{imprint} column using
a regex that identifies 4 numbers starting with 1 or 2 and extracts that as
an imputed date, after checking that it's not greater than the current year.
}
