% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_tools.R
\name{browse_htids}
\alias{browse_htids}
\title{Browse a set of Hathi Trust IDs interactively at the Hathi Trust digital
library site}
\usage{
browse_htids(htids)
}
\arguments{
\item{htids}{A workset generated by \link{workset_builder}, a character vector of
Hathi Trust IDs, or a data frame with a column named "htid" (and possibly
one named "id" as well for the page sequence).}
}
\value{
Opens a browser window at https://babel.hathitrust.org/ to the
specified Hathi Trust Id (and page, if given). If there is more than one
HTID in the workset, it will prompt you to open the next one.
}
\description{
Browse a set of Hathi Trust IDs interactively at the Hathi Trust digital
library site
}
\examples{
\donttest{
if(interactive()) {
   workset <- workset_builder(name = "Alexis de Tocqueville", pub_date = 1800:1850)
   browse_htids(workset)
}
}
}
