% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{cache_all}
\alias{cache_all}
\title{Caches all downloaded JSON Extracted Features files}
\usage{
cache_all(
  dir = getOption("hathiTools.ef.dir"),
  cache_type = getOption("hathiTools.cachetype"),
  keep_json = TRUE
)
}
\arguments{
\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_type}{Type of caching used. Defaults to
\code{getOption("hathiTools.cachetype")}, which is "csv.gz" on load. Allowed
cache types are: compressed csv (the default), "none" (no local caching of
JSON download; only JSON file kept), "rds", "feather", file (suitable for
use with \link{arrow}), or "text2vec.csv" (a csv suitable for use with the
package \href{https://cran.r-project.org/package=text2vec}{text2vec}).}

\item{keep_json}{Whether to keep the downloaded json files. Default is
\code{TRUE}; if false, it only keeps the local cached files (e.g., the csv
files). This can save space.}
}
\description{
It is useful to run this function after running \link{rsync_from_hathi}; this way,
you can cache all your slow-to-load JSON Extracted Features files to a faster
to load format (e.g., feather or csv).
}
