% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{download_hathi}
\alias{download_hathi}
\title{Downloads a Hathi Trust Extracted Features file for a given Hathi Trust id}
\usage{
download_hathi(htid, dir = "hathi-ef")
}
\arguments{
\item{htid}{The Hathi Trust id of the item whose extracted features files are
to be downloaded.}

\item{dir}{The directory to download the extracted features files. Defaults
to "hathi-ef"; the directory will be automatically created if it doesn't
already exist.}
}
\value{
Saves the json and csv versions of the extracted features.
}
\description{
Downloads a Hathi Trust Extracted Features file for a given Hathi Trust id
}
\examples{
\dontrun{

# Download the 1863 version of "Democracy in America" by Tocqueville
download_hathi("miun.aew4744.0001.001", dir = tempdir())
}
}
