% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{download_hathi_ef}
\alias{download_hathi_ef}
\title{Downloads a Hathi Trust Extracted Features file for a single Hathi Trust id}
\usage{
download_hathi_ef(
  htid,
  dir = getOption("hathiTools.ef.dir"),
  cache_type = getOption("hathiTools.cachetype")
)
}
\arguments{
\item{htid}{The Hathi Trust id of the item whose extracted features files are
to be downloaded.}

\item{dir}{The directory to download the extracted features files. Defaults
to \code{getOption("hathiTools.ef.dir")}, which is just "hathi-ef" on load; the
directory will be automatically created if it doesn't already exist.}

\item{cache_type}{Type of caching used. Defaults to
\code{getOption("hathiTools.cachetype")}, which is "csv.gz" on load. Allowed
cache types are: compressed csv (the default), "none" (no local caching of
JSON download; only JSON file kept), "rds", "feather", file (suitable for
use with \link{arrow}), or "text2vec.csv" (a csv suitable for use with the
package \link{text2vec}).}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the extracted features.
}
\description{
This function downloads a Hathi Trust Extracted Features file for a single
Hathi Trust id and (typically, unless explicitly prevented) caches the result
as a CSV file or other type of columnar format suitable for fast reading.
}
\details{
Note that if you want to download the Extracted Features of many Hathi Trust
IDs, it is best to use \link{rsync_from_hathi}. That is what you'd normally do if
you first built a workset using \link{workset_builder}.
}
\examples{
\dontrun{

# Download the 1862 version of "Democracy in America" by Tocqueville
download_hathi_ef("mdp.39015001796443", dir = tempdir())
}
}
