% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathifiles.R
\name{download_hathifile}
\alias{download_hathifile}
\title{Downloads the Hathi Trust big hathifile}
\usage{
download_hathifile(
  url = "https://www.hathitrust.org/hathifiles",
  dir = getOption("hathiTools.hathifile.dir"),
  full_catalog = TRUE
)
}
\arguments{
\item{url}{The URL for the Hathi Trust hathifiles
https://www.hathitrust.org/hathifiles}

\item{dir}{The directory to use to save the downloaded hathifile. Defaults to
\code{getOption("hathiTools.hathifile.dir")}, which on loading the package is
just \code{./raw-hathifiles} (a directory which will be created if it doesn't
exist already when you call the function).}

\item{full_catalog}{Whether to download the full catalog, or just the latest
update (there's a new "update file" every day, and a new version of the
full catalog every month). Default is \code{TRUE} - download the full catalog.}
}
\value{
The downloaded filename.
}
\description{
This function downloads the \href{https://www.hathitrust.org/hathifiles}{big hathifile catalog} with simple metadata for the
digitized volumes in the Hathi Trust digital library collection. It can be
used in conjunction with \link{workset_builder} and \code{rsync} to select an
appropriate sample of Hathi Trust Extracted Features files and metadata for
further analysis. Warning - it's a 1GB file; if the latest version of the
file (there's a new one every month) has been downloaded already, the
function will just return the file name and won't attempt to download it
again.
}
