% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{drama}
\alias{drama}
\title{Drama Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17709 rows and 3 columns.
}
\source{
\url{https://wiki.htrc.illinois.edu/display/COM/Word+Frequencies+in+English-Language+Literature,+1700-1922}
}
\usage{
drama
}
\description{
A dataset containing the HTIDs of the volumes classified as "drama" in Ted
Underwood, Boris Capitanu, Peter Organisciak, Sayan Bhattacharyya, Loretta
Auvil, Colleen Fallaw, J. Stephen Downie (2015). Word Frequencies in
English-Language Literature, 1700-1922 (0.2)
\href{http://data.analytics.hathitrust.org/genre/drama_metadata.csv}{Dataset}.
HathiTrust Research Center. doi:10.13012/J8JW8BSJ. Taken from the summary
netadata file at
http://data.analytics.hathitrust.org/genre/drama_metadata.csv
}
\details{
\describe{

\item{htid}{The Hathi Trust ID of the volume.}

\item{drama_prob}{A confidence metric: the probability that more than 80\% of
the pages in the volume assigned to "drama" have been correctly classified.}

\item{drama_prop}{The proportion of pages in the volume classified as
"drama". Calculated from \code{genrepages}/\code{totalpages} in the original metadata
file.}

}
}
\keyword{datasets}
