% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tools.R
\name{find_cached_htids}
\alias{find_cached_htids}
\title{Finds cached Extracted Features files for a set of HT ids}
\usage{
find_cached_htids(
  htids,
  dir = getOption("hathiTools.ef.dir"),
  cache_type = c("ef", "meta", "pagemeta"),
  cache_format = getOption("hathiTools.cacheformat"),
  existing_only = TRUE
)
}
\arguments{
\item{htids}{A character vector of Hathi Trust ids, a workset created with
\link{workset_builder}, or a data frame with a column named "htid" containing
the Hathi Trust ids that require caching.}

\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_type}{Type of information cached. The default is c("ef", "meta",
"pagemeta"), which refers to the extracted features, the volume metadata,
and the page metadata in \code{dir}. Omitting one of these caches or finds only
the rest (e.g., \code{cache_type = "ef"} caches only the EF files, not
their associated metadata or page metadata).}

\item{cache_format}{File format of cache for Extracted Features files.
Defaults to \code{getOption("hathiTools.cacheformat")}, which is "csv.gz" on
load. Allowed cache types are: compressed csv (the default), "none" (no
local caching of JSON download; only JSON file kept), "rds", "feather" and
"parquet" (suitable for use with \link{arrow}; needs the \link{arrow} package
installed), or "text2vec.csv" (a csv suitable for use with the package
\href{https://cran.r-project.org/package=text2vec}{text2vec}).}

\item{existing_only}{Whether to return only file paths to files that actually
exist. Default is \code{TRUE}. Use \code{FALSE} to find whether some files still need
to be cached.}
}
\value{
A \link{tibble} with the paths of the cached files and an indicator of
whether each htid has an existing cached file.
}
\description{
Finds cached Extracted Features files for a set of HT ids
}
