% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{get_hathi_counts}
\alias{get_hathi_counts}
\title{Reads the downloaded extracted features file for a given Hathi Trust id}
\usage{
get_hathi_counts(
  htid,
  dir = getOption("hathiTools.ef.dir"),
  cache_format = getOption("hathiTools.cacheformat")
)
}
\arguments{
\item{htid}{The Hathi Trust id of the item whose extracted features files are
to be loaded into memory. If it hasn't been downloaded, the function will
try to download it first.}

\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_format}{File format of cache for Extracted Features files.
Defaults to \code{getOption("hathiTools.cacheformat")}, which is "csv.gz" on
load. Allowed cache types are: compressed csv (the default), "none" (no
local caching of JSON download; only JSON file kept), "rds", "feather" and
"parquet" (suitable for use with \link{arrow}; needs the \link{arrow} package
installed), or "text2vec.csv" (a csv suitable for use with the package
\href{https://cran.r-project.org/package=text2vec}{text2vec}).}
}
\value{
a \link[tibble:tibble]{tibble} with the extracted features.
}
\description{
Given a single Hathi Trust ID, this function returns a
\link[tibble:tibble]{tibble} with its per-page word count and part of speech
information, and caches the results to the \code{getOption("hathiTools.ef.dir")}
directory (by default "./hathi-ef"). If the file has not been cached already,
it first attempts to download it directly from the Hathi Trust server. This
function uses code authored by Ben Schmidt, from his Hathidy package
(\url{https://github.com/HumanitiesDataAnalysis/hathidy}).
}
\examples{
\donttest{
# Download the 1863 version of "Democracy in America" by Tocqueville and get
# its extracted features

tmp <- tempdir()

get_hathi_counts("aeu.ark:/13960/t3qv43c3w", dir = tmp)

}
}
\author{
Ben Schmidt
}
