% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{get_hathi_meta}
\alias{get_hathi_meta}
\title{Reads the metadata of a single downloaded Hathi Trust extracted features file}
\usage{
get_hathi_meta(htid, dir = getOption("hathiTools.ef.dir"))
}
\arguments{
\item{htid}{The Hathi Trust id of the item whose extracted features files are
to be downloaded.}

\item{dir}{The directory where the file is saved. Defaults to
\code{getOption("hathiTools.ef.dir")}, which is just "./hathi-ef/" on load. If
the file does not exist, this function will first attempt to download it.}
}
\value{
a \link{tibble} with metadata.
}
\description{
Given a single Hathi Trust ID, this function returns a \link{tibble} with its
metadata information. If the HT EF file corresponding to this ID has not been
downloaded already, it first attempts to download it directly from the Hathi
Trust server. This function uses code authored by Ben Schmidt, from his
Hathidy package
(\href{https://github.com/HumanitiesDataAnalysis/hathidy}{https://github.com/bmschmidt/edinburgh/}).
}
\details{
Note that if you want to extract the metadata of more than one Hathi Trust ID
at a time, it is best to simply query the Workset Builder database using
\link{get_workset_meta}. Using \link{get_workset_meta} also guarantees you'll get a
rectangular data with HT ids in the rows. It is also possible to get simple
metadata for large numbers of htids by downloading the big hathifile using
\link{download_hathifile} and then filtering it.
}
\examples{
\dontrun{
# Download the 1862 version of "Democracy in America" by Tocqueville and get
# its metadata

tmp <- tempdir()

get_hathi_meta("mdp.39015001796443", dir = tmp)
}
}
\author{
Ben Schmidt
}
