% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{get_hathi_page_meta}
\alias{get_hathi_page_meta}
\title{Reads the page-level metadata of a single Hathi Trust Extracted Features file}
\usage{
get_hathi_page_meta(htid, dir = getOption("hathiTools.ef.dir"))
}
\arguments{
\item{htid}{The Hathi Trust id of the item whose extracted features files are
to be downloaded.}

\item{dir}{The directory where the file is saved. Defaults to
\code{getOption("hathiTools.ef.dir")}, which is just "./hathi-ef/" on load. If
the file does not exist, this function will first attempt to download it.}
}
\value{
A \link{tibble} with the page-level metadata for the corresponding Hathi
Trust ID. The page-level metadata contains the following fields (taken from
\url{https://wiki.htrc.illinois.edu/pages/viewpage.action?pageId=79069329}):

\describe{

\item{seq}{The sequence number of the page in the volume. Corresponds to
the digital object, so that the first scan in the volume is "00000001",
which may be the cover, a title page, or something else.}

\item{Description}{A hash of the page content used to compute the features
for the page. Volumes in HathiTrust may be updated to improve scan or OCR
quality or correct an issue, which would cause the text data to change,
and, if features are reprocessed, a new hash would result.}

\item{calculatedLanguage}{The most probable language of the text on the
page. Determined algorithmically, and specified by language codes. Will be
\code{NA} if no language detected, or if the language was not recognized by the
algorithm.}

\item{tokenCount}{The total number of tokens detected on the page.}

\item{lineCount}{The total number of lines of text detected on the page.}

\item{emptyLineCount}{The total number of empty lines on the page.}

\item{sentenceCount}{The total number of sentences detected on the page.}

\item{sectionStats}{A \link{tibble} column with stats for each of the sections
of the page ("header", "body", and "footer"). This tibble has the following
fields:
\itemize{
\item section: The seciton of the page.
\item sectiontokenCount: The total number of tokens detected in the section of
the page.
\item sectionlineCount: The total number of lines detected in the section of
the page.
\item sectionemptyLineCount The total number of empty lines detected in the
section of the page.
\item sectionsentenceCount The total number of sentences detected in the
section of the page.
\item sectioncapAlphaSeq The longest length of the alphabetical sequence of
capital characters starting a line. Only available for the "body" section.
\item beginCharCount A \link{tibble} column with the first non-White Space
characters detected on lines in the section (beginChar) and their
occurrence counts (beginCharCount).
\item endCharCount An \link{tibble} column of the last non-White Space characters on
detected lines in the section (endChar) and their occurrence counts
(endCharCount). }
}

}
}
\description{
Given a single Hathi Trust ID, this function returns a
\link[tibble:tibble]{tibble} with its page-level metadata information. If the HT
EF file corresponding to this ID has not been downloaded already, it first
attempts to download it directly from the Hathi Trust server. This function
uses code authored by Ben Schmidt, from his Hathidy package
(\url{https://github.com/HumanitiesDataAnalysis/hathidy}).
}
\details{
Note that if you want to extract the page-level metadata of more than one
Hathi Trust ID at a time, it may be best to download the JSON files for these
HTIDs first using \link{rsync_from_hathi} and then running this function.
}
\examples{
\dontrun{
# Download the 1862 version of "Democracy in America" by Tocqueville and get
# its page-level metadata

tmp <- tempdir()

page_meta <- get_hathi_page_meta("mdp.39015001796443", dir = tmp)

page_meta \%>\% tidyr::unnest(sectionStats)
page_meta \%>\% tidyr::unnest(sectionStats) \%>\% tidyr::unnest(beginCharCount)

}
}
\author{
Ben Schmidt

Xavier Marquez
}
