% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync_tools.R
\name{htid_to_rsync}
\alias{htid_to_rsync}
\title{Converts a list of htids to relative paths for rsync to download}
\usage{
htid_to_rsync(htids, file)
}
\arguments{
\item{htids}{A character vector or list of HathiTrust ids (htids), a workset
generated by \link{workset_builder}, or a data frame with a column named 'htid'
and containing the htids.}

\item{file}{A text file to save the resulting list of relative stubbytree
paths to use in the command \verb{rsync -av --files-from FILE.txt data.analytics.hathitrust.org::features-2020.03/ hathi-ef/}}
}
\value{
The list of relative paths saved to the file (invisibly).
}
\description{
Converts a list of htids to relative paths for rsync to download
}
\section{Details}{


If you have a lot of files to download, generating the list of relative
stubbytree paths and using rsync is much faster than using
\link{get_hathi_counts} over a list of htids. But rsync only downloads json
files, so calling \link{get_hathi_counts} on a downloaded json file will be
slower the first time as the function will cache the json file to csv. It
is best to run \link{cache_htids} after using rsync to reduce this performance
penalty.
}

\examples{
htid_to_rsync(c("nc01.ark:/13960/t2v41mn4r", "mdp.39015001796443"), tempfile())
}
