% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{iso639}
\alias{iso639}
\title{ISO639 language codes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 487 rows and 5 columns.
}
\source{
\url{https://datahub.io/core/language-codes#resource-language-codes-full}
}
\usage{
iso639
}
\description{
A dataset containing the ISO639 language codes. Used internally to match
languages.
}
\details{
\describe{

\item{alpha3-b}{alpha-3 bibliographic code}

\item{alpha3-t}{alpha-3 terminologic code (when given)}

\item{alpha2}{alpha-2 code (when given)}

\item{English}{English name of language}

\item{French}{French name of language}
}
}
\keyword{datasets}
