% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathifiles.R
\name{load_raw_hathifile}
\alias{load_raw_hathifile}
\title{Loads the raw hathifile into memory}
\usage{
load_raw_hathifile(
  filename = NULL,
  dir = getOption("hathiTools.hathifile.dir"),
  hathifile_date = getOption("hathiTools.hathifile"),
  cols,
  fix_date = TRUE
)
}
\arguments{
\item{filename}{The name of the downloaded hathifile. If \code{NULL}, it will
attempt to guess it from \code{getOption("hathiTools.hathifile.dir")} and
\code{getOption("hathiTools.hathifile")}; if it can't find it or the file doesn't
exist, it will attempt to download it to the directory in
\code{getOption("hathiTools.hathifile.dir")} using \link{download_hathifile}.}

\item{dir}{The directory where the raw hathifile is to be found.}

\item{hathifile_date}{The date of the hathifile. (A new one is released every
month). This defaults to \code{getOption("hathiTools.hathifile")}, which is just
the date of the last downloaded hathifile.}

\item{cols}{If wanted, a set of columns to load. Since the file is so large,
one can reduce memory use by selecting only certain columns. These can be
any of the following: \verb{htid (required), access, rights, ht_bib_key,}
\verb{description, source, source_bib_num, oclc_num, isbn, issn, lccn, title,}
\verb{imprint, rights_reason_code, rights_timestamp, us_gov_doc_flag,}
\verb{rights_date_used, pub_place, lang, bib_fmt, collection_code,}
\verb{content_provider_code, responsible_entity_code, digitization_agent_code,}
\verb{access_profile_code, author}. If cols = "REDUCED", the function loads a
reduced set of columns: \verb{htid, ht_bib_key, description, source, title,}
\verb{imprint, rights_date_used, us_gov_doc_flag}, \code{lang}, \code{bib_fmt}, and
\code{author}}

\item{fix_date}{Fixes \code{9999} values in \code{rights_date_used} by changing them to
\code{NA}. Default is \code{TRUE}.}
}
\value{
A very large \link{tibble}, with over 17 million records, loaded into
memory. The tibble package does some lazy loading to minimize resource use,
but fully loaded this data frame takes over 5GB in memory.
}
\description{
This function loads a previously downloaded hathifile into memory (or
downloads the latest one if it can't find it). It also turns the column
\code{us_gov_doc_flag} into a logical and eliminates \code{9999} values for
\code{rights_date_used}.
}
