% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_api.R
\name{query_ht_metadata_api}
\alias{query_ht_metadata_api}
\title{Queries the Hathi Trust Catalog API}
\usage{
query_ht_metadata_api(htid, type = c("brief", "full"))
}
\arguments{
\item{htid}{The Hathi Trust ID to query}

\item{type}{The type of query; the default is "brief". If using "full", the
response will contain the full MARC catalog record (as unparsed XML; you'll
have to do further processing to extract its data).}
}
\value{
A list created from the JSON response with the metadata information,
if it exists.
}
\description{
This is useful if one wants to check the metadata for a single Hathi Trust ID
quickly without downloading the whole text. It will not work for multiple
HTIDs.
}
\examples{
\dontrun{
query_ht_metadata_api("ien.35556036918548")
query_ht_metadata_api("ien.35556036918548", type = "full")
}
}
