% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tools.R
\name{read_cached_htids}
\alias{read_cached_htids}
\title{Read Cached HTIDs}
\usage{
read_cached_htids(
  htids,
  dir = getOption("hathiTools.ef.dir"),
  cache_type = getOption("hathiTools.cachetype"),
  include_meta = TRUE,
  include_page_meta = TRUE,
  filter_expression = TRUE
)
}
\arguments{
\item{htids}{A character vector of Hathi Trust ids, a workset created with
\link{workset_builder}, or a data frame with a column named "htid" containing
the Hathi Trust ids that require caching.}

\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_type}{Type of caching used. Defaults to
\code{getOption("hathiTools.cachetype")}, which is "csv.gz" on load. Allowed
cache types are: compressed csv (the default), "none" (no local caching of
JSON download; only JSON file kept), "rds", "feather" (suitable for use
with \link{arrow}; needs the \link{arrow} package installed), or "text2vec.csv" (a
csv suitable for use with the package
\href{https://cran.r-project.org/package=text2vec}{text2vec}).}

\item{include_meta}{Whether to include the volume-level metadata in the
returned \link[tibble:tibble]{tibble}. Default is \code{TRUE}.}

\item{include_page_meta}{Whether to include the page-level metadata in the
returned \link[tibble:tibble]{tibble}. Default is \code{TRUE}.}

\item{filter_expression}{A filter expression to filter each returned
\link[tibble:tibble]{tibble} by (so as to only load into memory desired
features).}
}
\value{
A \link[tibble:tibble]{tibble} with the extracted features, plus the
desired (volume-level or page-level) metadata.
}
\description{
Read Cached HTIDs
}
