% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync_tools.R
\name{rsync_from_hathi}
\alias{rsync_from_hathi}
\title{rsync Hathi Trust EFs from Hathi Trust}
\usage{
rsync_from_hathi(htids, dir = getOption("hathiTools.ef.dir"))
}
\arguments{
\item{htids}{A vector of Hathi Trust IDs, a workset generated by
\link{workset_builder}, or a data frame with a column named 'htid'.}

\item{dir}{The directory to save the downloaded EF files to. This defaults to
\code{getOption("hathiTools.ef.dir")}, which is just  "./hathi-ef" on loading.}
}
\value{
The return code of \link{system} or \link{system2} used to run \code{rsync}, which
must be installed on your system.
}
\description{
Given a vector of Hathi Trust IDs (generated, for example, from
\link{workset_builder}), this function attempts to use
\href{https://en.wikipedia.org/wiki/Rsync}{rsync} to download their Extracted
Features (per-page word counts and part of speech information) files from
Hathi Trust. \code{rsync} needs to be installed in your system and accessible via
\link{system} or \link{system2}.
}
\examples{
\dontrun{
res <- workset_builder(c("tylenol", "paracetamol"), volumes_only = FALSE)
rsync_from_hathi(res$htid[1:10])
}
}
