% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tools.R
\name{cache_htids}
\alias{cache_htids}
\title{Caches downloaded JSON Extracted Features files to another format}
\usage{
cache_htids(
  htids,
  dir = getOption("hathiTools.ef.dir"),
  cache_type = c("ef", "meta", "pagemeta"),
  cache_format = getOption("hathiTools.cacheformat"),
  keep_json = TRUE
)
}
\arguments{
\item{htids}{A character vector of Hathi Trust ids, a workset created with
\link{workset_builder}, or a data frame with a column named "htid" containing
the Hathi Trust ids that require caching.}

\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_type}{Type of information cached. The default is c("ef", "meta",
"pagemeta"), which refers to the extracted features, the volume metadata,
and the page metadata in \code{dir}. Omitting one of these caches or finds only
the rest (e.g., cache_type = "ef" caches or finds only the EF files, not
their associated metadata or page metadata).}

\item{cache_format}{File format of cache for Extracted Features files.
Defaults to \code{getOption("hathiTools.cacheformat")}, which is "csv.gz" on
load. Allowed cache types are: compressed csv (the default), "none" (no
local caching of JSON download; only JSON file kept), "rds", "feather"
(suitable for use with \link{arrow}; needs the \link{arrow} package installed), or
"text2vec.csv" (a csv suitable for use with the package
\href{https://cran.r-project.org/package=text2vec}{text2vec}).}

\item{keep_json}{Whether to keep the downloaded json files. Default is
\code{TRUE}; if \code{FALSE}, it only keeps the local cached files (e.g., the csv
files) and deletes the associated JSON files.}
}
\value{
A \link{tibble} with the paths of the cached files and an indicator of
whether each htid was successfully cached.
}
\description{
It is useful to run this function after running \link{rsync_from_hathi}; this way,
you can cache all your slow-to-load JSON Extracted Features files to a faster
to load format (e.g., \code{feather} or \code{csv}). By default, this function also
caches the associated metadata as separate .rds files.
}
