% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tools.R
\name{clear_cache}
\alias{clear_cache}
\title{Removes cached Extracted Features files for a set of HT ids}
\usage{
clear_cache(
  htids,
  dir = getOption("hathiTools.ef.dir"),
  cache_type = c(getOption("hathiTools.cachetype"), "meta", "pagemeta"),
  keep_json = TRUE
)
}
\arguments{
\item{htids}{A character vector of Hathi Trust ids, a workset created with
\link{workset_builder}, or a data frame with a column named "htid" containing
the Hathi Trust ids that require caching.}

\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_type}{The type of cached file to remove. Defaults to
\code{getOption("hathiTools.cachetype")}, which is "csv.gz" on load, as well as
any cached volume-level or page-level metadata ("meta" and "pagemeta").
Other possibilities are: "rds", "feather" (suitable for use with \link{arrow};
needs the \link{arrow} package installed), or "text2vec.csv" (a csv suitable for
use with the package
\href{https://cran.r-project.org/package=text2vec}{text2vec})}

\item{keep_json}{Whether to keep any downloaded JSON files. Default is
\code{TRUE}; if \code{FALSE} will delete all JSON extracted features associated with
the set of htids.}
}
\value{
A \link{tibble} with the paths of the cached files and an indicator of
whether each htid has an existing cached file.
}
\description{
Removes cached Extracted Features files for a set of HT ids
}
