% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{get_hathi_counts}
\alias{get_hathi_counts}
\title{Reads the downloaded extracted features file for a given Hathi Trust id}
\usage{
get_hathi_counts(htid, dir = "hathi-ef")
}
\arguments{
\item{htid}{The Hathi Trust id of the file whose extracted features are to be
returned. If the file does not exist, it is first downloaded using \link{download_hathi}.}

\item{dir}{The directory where the file is saved. Defaults to \code{hathi-ef}.}
}
\value{
a \link{tibble} with the extracted features, by page.
}
\description{
Reads the downloaded extracted features file for a given Hathi Trust id
}
\examples{
\dontrun{
# Download the 1863 version of "Democracy in America" by Tocqueville

tmp <- tempdir()
download_hathi("miun.aew4744.0001.001", dir = tmp)

get_hathi_counts("miun.aew4744.0001.001", dir = tmp)

}
}
