% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-ef-tools.R
\name{get_hathi_counts}
\alias{get_hathi_counts}
\title{Reads the downloaded extracted features file for a given Hathi Trust id}
\usage{
get_hathi_counts(
  htid,
  dir = getOption("hathiTools.ef.dir"),
  cache_type = getOption("hathiTools.cachetype")
)
}
\arguments{
\item{htid}{The Hathi Trust id of the item whose extracted features files are
to be loaded into memory. If it hasn't been downloaded, the function will
try to download it first via \link{download_hathi_ef}.}

\item{dir}{The directory where the download extracted features files are to
be found. Defaults to \code{getOption("hathiTools.ef.dir")}, which is just
"hathi-ef" on load.}

\item{cache_type}{Type of caching used. Defaults to
\code{getOption("hathiTools.cachetype")}, which is "csv.gz" on load. Allowed
cache types are: compressed csv (the default), "none" (no local caching of
JSON download; only JSON file kept), "rds", "feather", file (suitable for
use with \link{arrow}), or "text2vec.csv" (a csv suitable for use with the
package \href{https://cran.r-project.org/package=text2vec}{text2vec}).}
}
\value{
a \link{tibble} with the extracted features.
}
\description{
Reads the downloaded extracted features file for a given Hathi Trust id
}
\examples{
\dontrun{
# Download the 1863 version of "Democracy in America" by Tocqueville

tmp <- tempdir()

download_hathi_ef("aeu.ark:/13960/t3qv43c3w", dir = tmp)

#Get the downloaded extracted features
get_hathi_counts("aeu.ark:/13960/t3qv43c3w", dir = tmp)

}
}
