% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workset_builder.R
\name{get_workset_meta}
\alias{get_workset_meta}
\title{Get metadata for a set of Hathi Trust IDs}
\usage{
get_workset_meta(
  workset,
  metadata_dir = getOption("hathiTools.metadata.dir"),
  cache = TRUE
)
}
\arguments{
\item{workset}{A workset of htids, generated by \link{workset_builder} from \href{https://solr2.htrc.illinois.edu/solr-ef/}{Hathi Trust's SOLR endpoint}. One can
also pass a data frame with a column labeled "htid" and containing valid
Hathi Trust htids, or a character vector of htids (though the function will
complain with a warning).}

\item{metadata_dir}{The directory used to cache the metadata file. Defaults
to \code{getOption("hathiTools.metadata.dir")}, which is just "./metadata" on
loading the package.}

\item{cache}{Whether to cache the resulting metadata as a CSV. Default is
TRUE. The name of the resulting metadata file is generated by appending an
MD5 hash (via \link[digest:digest]{digest::digest}) to the string "metadata-", so each metadata
download will have a different name.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the Hathi Trust metadata for all the volumes
in the workset or the vector of htids.
}
\description{
Queries the SOLR endpoint of the Workset Builder 2.0 (beta) at
https://solr2.htrc.illinois.edu/solr-ef/ to download volume metadata. This
API is experimental, and so this function can fail at any time if the API
changes.
}
\details{
Be mindful that downloading a large number of metadata records can take
quite some time. In practice I have found that downloading full metadata from
more than about 1000 records is a dicey proposition; if you need metadata for
many thousands of records, you are probably better off using the big
hathifile (see \link{download_hathifile} and \link{load_raw_hathifile}).
}
\examples{
\dontrun{
workset <- workset_builder(name = "Tocqueville")
meta <- get_workset_meta(workset)

## We can also pass a vector of htids:
meta <- get_workset_meta(workset$htid)
}
}
